fix: don't assume files to have CR characters on Windows
The current code assumes that all lines of a JavaScript file end with CR+LF on Windows. However, for example, if the `eol` attribute of the project's .gitattributes is "lf", files don't have CR characters even on Windows:

```console
$ node -p "fs.readFileSync('1.js')" # when .gitattribute is "* text=auto"
<Buffer 31 0d 0a>

$ node -p "fs.readFileSync('1.js')" # when .gitattribute is "* text=auto eol=lf"
<Buffer 31 0a>
```

Because of this assumption, currently v8-to-istanbul cannot calculate line coverage correctly especially on Windows.

This patch solves the problem above.