bug: Empty sources and offset issue
This PR corrects two bugs:

* For source maps, the sources property can potentially be an empty array if mappings
doesn't provide any information about its origin. For example if you compile an empty
js file with babel, a valid sourcemap can be created with `[]` for `sources`. Here
we add a fallback to `file` in this case and adds a fixture to test for it.

* When a file has dos line endings `\r\n` we found a few instances in a large codebase
where v8 would report the position of the `\n` as `startCol` for a range. This would
cause the starting column supplied to `sourceMap.originalPositionFor` to be a `-1`
because the startColumns of the transpiled source lines were calculated as
`endCol + length of line break sequence`. This change clamps the supplied value to
max 0 and adds a test fixture (force encoded as `\r\n`) to test for it.