fix: Handle sourcemap `sources` emtpy edge case
For source maps, the sources property can potentially be an empty array if mappings
doesn't provide any information about its origin. For example if you compile an empty
js file with babel, a valid sourcemap can be created with `[]` for `sources`. Here
we add a fallback to `file` in this case and adds a fixture to test for it.

The new test fixtures `empty.js`, `emtpy.compiled.js` and `empty.compiled.js.map`
simulate this scenario for a unit test.