fix: v8 coverage ranges that fall on \n characters cause exceptions (#96)
When a file has dos line endings \r\n we found a few instances in a large codebase
where v8 would report the position of the \n as startCol for a range. This would
cause the starting column supplied to sourceMap.originalPositionFor to be a -1
because the startColumns of the transpiled source lines were calculated as
endCol + length of line break sequence.

To replicate this condition, revert the fix but run the new unit test. The
call to `applyCoverage` will crash. While the unit test range was hand
crafted by me for this particular example, we saw this coming up in the
wild.

This change clamps the supplied value to
max 0 and adds a test fixture (force encoded as \r\n) to test for it.

Co-authored-by: Joe Mordetsky <jmordetsky@bloomberg.net>