fix: fix undefined line in branches and functions
When bundling this [file](https://github.com/browserify/node-util/blob/master/util.js) with esbuild and getting the v8 coverage with playwright `_branchesToIstanbul` and `_functionsToIstanbul` hanged.

These 2 functions actually should throw an error but it is swallowed for some reason.

This PR just ignores if it can't find the source line for the current branch or function.